feat: estrutura inicial da API com rotas de usuário e configuração do ESLint
