+++ packages/api-node/app.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import userRoutes from './routes/userRoutes.js';
+
+const app = express();
+app.use(express.json());
+app.use('/users', userRoutes);
+
+export default app;
