+++ packages/api-node/routes/userRoutes.js
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import { getUsers, createUser } from '../controllers/userController.js';
+
+const router = Router();
+
+router.get('/getUsers', getUsers);
+router.post('/createUser', createUser);
+
+export default router;
